+++ GradeDataModel/StudentGrade.cs
@@ -3,16 +3,16 @@
     public class StudentGrade
     {
         public string StudentName { get; set; } = string.Empty;
-        public double Quiz1 { get; set; }
-        public double Quiz2 { get; set; }
+        public decimal Quiz1 { get; set; }
+        public decimal Quiz2 { get; set; }
 
-        public double Quiz3 { get; set; }
-        public double LongQuiz1 { get; set; }
-        public double LongQuiz2 { get; set; }
-        public double Project { get; set; }
-        public double Perform { get; set; }
-        public double Mid { get; set; }
-        public double Finals { get; set; }
-        public double TotalGrade { get; set; }
+        public decimal Quiz3 { get; set; }
+        public decimal LongQuiz1 { get; set; }
+        public decimal LongQuiz2 { get; set; }
+        public decimal Project { get; set; }
+        public decimal Perform { get; set; }
+        public decimal Mid { get; set; }
+        public decimal Finals { get; set; }
+        public decimal TotalGrade { get; set; }
     }
 }
